Moved XNAT tools to /shared rather than only classifier 2d
Also udpated nomenclature for xnat 0.8
